[WFARQ-71] Create a common deployable container for the bootable container and managed container to share.
